Use non-recursive shared lock for Manifest::rwlock
Currently VB::Manifest::rwlock uses cb::RWLock for the shared
lock. This is implemented using pthread_rwlock underneath, which is
(implicitly) a recursive mutex.

We don't actually require a recursive mutex for VB::Manifest::rwlock
(a thread shouldn't attempt to read-lock it twice), so to avoid any
potential bugs where we lock too many times change to use a
non-recursive mutex.

The chosen mutex is (initially) std::shared_mutex given it's the
simplest to drop in instead of RWLock. However we could consider using
folly::SharedMutex instead which is expected to be more performant
when being locked in shared mode (as is the common case for
VB::Manifest).

Change-Id: I24c743d0a20aa316ab7176b4319976fa6b4dd08b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/111236
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>