MB-34682: Allow abort without prepare if receiving disk snapshot
The previous attempt to allow aborts without a prepare limited the
valid window for this to aborts with a prepare seqno less than the
first seqno seen after stream resumption. This fails to account for
aborts overwriting prepares on disk so we will never see a prepare
in a disk snapshot if we will also see an abort.

Instead, we should just allow (and ignore) an abort without a prepare
if we are receiving a disk snapshot and the prepare seqno of the
abort is greater than or equal to the snapshot start; it should never
be less because we should have previously received it.

Change-Id: Icdb5aa9ebf7f2a9a212808a6d1116e9189577ea6
Reviewed-on: http://review.couchbase.org/111249
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>