MB-34751: "PLAIN" should be returned for ext auth
If the system is configured with external authentication through
ns_server (to access users in LDAP) the clients must use PLAIN
as the authentication mechanism.

A well written client would start by connecting to the server
to fetch the list of supported mechanisms to use to authenticate
to the server, and then pick one of the supported mechanisms.
If the server returned PLAIN,SCRAM-SHA1 the client would pick
SCRAM-SHA1 over PLAIN as it is more secure, but now the client
would fail to authenticate as the users in LDAP. In previous
versions of CB we required the client to be configured to
always use PLAIN iff they wanted to use a client defined
in LDAP.

Change-Id: I8606b4392e552f2dbb857c8fd30c118b09d5f498
Reviewed-on: http://review.couchbase.org/111289
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>