MB-34092: Correctly track FE doc count with Sync Writes
Currently the full eviction doc counting is incorrecet when Sync
Writes are used. The prepares on disk are counted in the number
of items reported for the vBucket. Correct this by only counting
committed items and remove the number of in flight prepares at
warmup from the total count loaded from disk.

Change-Id: I400d973c94bb94eba01d3006eb421aff73872272
Reviewed-on: http://review.couchbase.org/111296
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>