MB-34768: Only check committed items for expiry
Problem:
The compaction expiry hook tries to expire sync delete prepares.

Cause:
The expiry field is used to indicate deletion time for deletes, and is
set to the current time when processing the op. This is also applied to
sync delete prepares, but as they are not actually deletes,
couch-kvstore misinterprets them as an expired item.

Solution:
Ensure items are not candidates for expiry unless they are committed.
Prepares are never eligible for expiry, so this is valid.

Change-Id: Ic25b7fa44fcd6c50666e1db80797ac80004899c2
Reviewed-on: http://review.couchbase.org/111347
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>