MB-33463: Add support for delVBucket and reset API calls
As part of this checkin, we've added a MagmaKVHandle which protects
magma API calls from the vbucket being dropped. The MagmaKVHandle is a
shared_ptr which is acquired at the beginning of many of the API calls.
delVBucket can't proceed in dropping the vbucket until the use_count on
the handle becomes unique().

Inside the handle is the vbstateMutex. This mutex is used to control
race conditions between updaters of the vbstate and readers of the
vbstate.

Change-Id: I9a040cd1a81e4286464891393839bd5133314a81
Reviewed-on: http://review.couchbase.org/110928
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>