MB-34682: Clean up assertions for aborts with deduped prepares
Split the error checking and returning of ENGINE_EINVAL when we
receive an unexpected abort into two logically different
components for easier debugging with logs.

Rename getLatestSnapshotStartSeqno to getOpenSnapshotStartSeqno.

Add additional assertions:

If we see an abort with a prepare seqno that does not match the
prepare in the HashTable on the replica node then we expect to be
in disk backfill and the prepare seqno to be greater than or equal
to the snapshot start seqno (otherwise we must have seen it).


Change-Id: I7fd49b5fe5d9fdf3e854caa2d0baae7781630d23
Reviewed-on: http://review.couchbase.org/111410
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>