MB-34092: Persist the number of on disk prepares
*** PROBLEM ***

We need to have a correct onDiskTotalItems for each vBucket post-
warmup. This counter is warmed up from couchstore's number of items.
This number always includes the number of prepares on disk.

We need to persist a counter somewhere so that we can subract the
number of prepares on disk from the number of items on disk returned
by couchstore at warmup.

*** SOLUTION 1 ***

We could leverage the collection document counting code given that
all prepares on disk are persisted with a key with collection CID:2
and nothing else is.

This is not ideal because we would need to create a ManifestEntry
with CID:2 for every vBucket. This would cause an extra SystemEvent to
be persisted which in turn causes an extra DCP System Event to be sent
to any collections aware DCP clients. The number of tests that would
have to be updated for this is very high. Additionally we would waste
space in memory storing start and end seqnos and high seqnos that would
be duplicated if we didn't rework the HPS code. If we were to rework
the HPS code then we'd have a discrepancy between how we store the HPS
and the HCS.

*** SOLUTION 2 ***

Add a new field to vbstate. This is far easier to implement and the
field directly relates to the HCS and HPS values that we persist.

*** IMPLEMENTATION ***

See solution 2.

Change-Id: I101a349ef0102cdd766f842ca1131561963ec316
Reviewed-on: http://review.couchbase.org/111354
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>