MB-34810: Delete correct HashTable item from SyncDelete persistence callback
When a SyncDelete commit is persisted to disk is will trigger the
ep-engine persistence callback. As part of this, the now-committed
delete should be removed from the HashTable.

However, when removing the item from the HashTable via
{{HashTable::unlocked_del}}, the wrong overload is used (key instead
of StoredValue), which can result in a different, _prepared_
StoredValue from being removed instead.

Fix by using the overload which takes a StoredValue.

Change-Id: Ia844503019a57569924a89896c067678314b4b30
Reviewed-on: http://review.couchbase.org/111434
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>