MB-34789: Send mutation instead of commit for backfilled commits
Send a DCP Mutation instead of a DCP Commit when sending a commit
with prepare seqno > requested stream start seqno if the commit
has been backfilled from disk. We need to do this because the
de-duplication of keys on disk can result in a commit being the
first seqno a replica will see. In this case we must send a mutation
because the consumer needs to know what to commit. Currently, the
code does this if the prepare seqno is less than the requested
start seqno. This does not solve the case where we de-dupe a prepare
within the disk snapshot though. Instead, send the mutation instead
of the commit if the replica did not receive the prepare before it
connected whilst we are backfilling from disk.

*** EXAMPLE ***

XXXX = Nothing sent/received

Active:  P:k1            C:k1 M:k2 P:k1 C:k1 P:k1
Replica: P:k1 Disconnect XXXX M:k2 XXXX C:k1 P:k1

Replica disconnects and then streams from disk.

For the first commit the replica receives we must send a
mutation so that we can commit the correct value. The previous
code would send a commit as the prepare seqno was greater than the
first seqno sent.

Change-Id: I25bb8a11a920c035bfd25f656320642d4aba751e
Reviewed-on: http://review.couchbase.org/111435
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>