MB-34814: Make active send MARKER_FLAG_CHK with disk snapshots
Replicas should be instructed to open a new checkpoint if they are
receiving a disk snapshot in case they they encounter items
which might lead to non-permitted deduping (e.g., durability ops).

The active will never have queued such items into the same checkpoint,
but it didn't previously indicate to the replica that a new checkpoint
should be opened when sending from disk.

Change-Id: I4c76ecfb13afce14f42259d06b28d56797bce896
Reviewed-on: http://review.couchbase.org/111444
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>