Fix logging of connections during bucket deletion
During bucket deletion we don't want to constantly log all
of the connections, instead we want to log the connections
which changed properties since the last time we logged
them.

There was a bug in the previous code where we used the
previous diff to compare the current status, which caused
entries to reappear at a later time even if they don't
change.

Change-Id: Ic628e2bae8091be0a3f61da652c5f979727eb153
Reviewed-on: http://review.couchbase.org/111445
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>