MB-34815 [SR]: Remove HashTable::unlocked_del(Dockey&) overload
HashTable::unlocked_del() has an overload which takes a DocKey&. With
the advent of SyncWrites this method is ambiguous; it simply removes
the first StoredValue encountered with the given key which could
either be a Committed or Prepared item.

I see no use-case where the caller doesn't care which type of item to
remove - indeed all existing usages of this method were calling
HashTable::find...() methods to locate the particular type of
StoredValue they wanted to delete; but went on to call the DocKey&
which could end up removing a different StoredValue.

Replace all uses with the StoredValue* overload which is unambiguous.

Change-Id: Ia66a0933f361b5a03926f92dffeefbeec76d6dcd
Reviewed-on: http://review.couchbase.org/111448
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>