MB-34771: Abort in-flight SyncWrites if durability impossible
At topology change, we should abort in-flight SyncWrites if
durability is no longer possible. We should abort them instead of
returning E_DURABILITY_IMPOSSIBLE just in case some further
failover situation results in a commit.

To abort the SyncWrites, we must temporarily allow them to exist
even though durability is not possible. This is done under the
ADM::State write lock so no other thread will see this. We must do
this because they would otherwise have invalid chain pointers which
would result in segfaults when attempting to abort them.

Change-Id: I2124a63a395a66331c7be543f91ca4dffbc11ae5
Reviewed-on: http://review.couchbase.org/111486
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>