Add labels to {Atomic,Weakly,}Monotonic
Add support for a label to be associated with the various Monotonic
template class instantiations. This aids in reporting of monotonic
invariants being broken - previously there was little context on
exactly what had failed. Example output:

    Monotonic<x> (ActiveDM(vb:0)::State::lastAbortedSeqno) invariant failed: new value (302) breaks invariant on current value (303)

Note this increases the size of Monotonic from sizeof(T) to ~
sizeof(T) + 24. Where Monotonic was used where a specific size was
relied on, updates usages to use Monotonic::value_type which is the
underlying templated type.

Change-Id: I1a9445ab24d9136361537fd1bfd9e100406b1a81
Reviewed-on: http://review.couchbase.org/111506
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>