MB-34778: Test log flushing thread destruction of BucketLogger
The destruction of a BucketLogger calls
cb::logger::unregister_logger to remove the logger from the spdlog
registry. This caused a deadlock in spdlog when the background
flushing thread was the only owner of the shared_ptr. It would
attempt to destruct the BucketLogger but deadlock as it already held
the logger_map_mutex_ in spdlog::registry. Spdlog registry has been
updated to make this mutex recursive which will allow the double
locking by this thread.

Change-Id: Ie9590c6161f5cd391ccd420023f9311a602e87e7
Reviewed-on: http://review.couchbase.org/111489
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>