MB-34813: Ignore abort if prepare deduped and last committed
If a replica disconnects and reconnects then it may receive the
following sequence of events.

XXXX = Nothing sent/received

Active:  M:k1            P:k1 A:k1
Replica: M:k1 Disconnect XXXX A:k1

Currently we throw an exception as the Abort finds a non-pending
StoredValue. We should ignore the Abort as the prepare has been
de-duped.

Change-Id: Ie7afc32b3fccf851f591ea56101ea6d2c022be59
Reviewed-on: http://review.couchbase.org/111529
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>