MB-34848: Initialize ADM lastTrackedSeqno at replica promotion
The lastTrackedSeqno in the Active DM is not initialized by default
and will not be initialized by a replica promoted to active if
the are no tracked writes. If there are no tracked writes then it
should be set to the high completed seqno as we must have completed
all writes (if we had any).

Change-Id: I4ed7fb77a14df51fe25fe7c941d0cb8ab56d8ff0
Reviewed-on: http://review.couchbase.org/111555
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>