MB-34718: 3/3 Introduce new warmup step that adds vbuckets to the vbMap
The bug in MB-34718 is in full-eviction we can process a GET whilst
warmup is running *and* go on to flush a delete (if the GET found an
expired item). The flush then attempts to decrement the collection
counter which is still in the default initialised state of 0. -1 is not
allowed and an exception takes down KV-engine.

This commit addresses the bug seen by only adding the warmed up VBuckets
to the vbMap once they've been fully initialised from all disk state.

This includes:
* VB state loaded
* collection manifest loaded
* collection count loaded (as per the bug)
* bucket item count
* prepared items

Once all of these items are loaded into the VBucket, the VBucket can
be made available to the rest of KV-engine.

Change-Id: I0bfff240abcf436cc2f21f6f887a765a89698eb0
Reviewed-on: http://review.couchbase.org/111559
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>