Cleanup SW stream backfill tests
Currently the tests do not run on Ephemeral as they attempt to
backfill a prepare with abort of MajorityAndPersistToMaster level.
Ephemeral supports Majority level only and will backfill all
prepares as Majority. As such, we can remove the old todos and
test Majority level on Ephemeral and Persistent buckets and
test MajorityAndPersistToMaster and PersistToMajority on Persitent
buckets only.

Also, test the backfilled prepare seqno of an abort.

Change-Id: I086356d8ffe35bdb5a8068847dd9f8699e4f65f6
Reviewed-on: http://review.couchbase.org/111564
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>