[cluster_test] Add support for DCP packet filter
Add support for registering a packet filter to the DCP
streams. It may inspect, modify (or even drop) the packet.

The motivation for this is to be able to monitor (or mock)
the state of our system (for instance detect when a DCP Prepare
is sent to a certain node, and possibly drop it)

Change-Id: I4b58ef49809ea8d5bae40508528e102d83cb54e5
Reviewed-on: http://review.couchbase.org/111584
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>