[SR] Accept respones to Sync-Repl DCP messages and pass to engine
The Sync Replication DCP opcodes DcpPrepare, DcpCommit & DcpAbort are
not currently handled in memcached. This means that if a (producer)
memcached receives these messages they are rejected at the memcached
(and just disconnect the connection) without being passed down to the
engine to decide what to do with them.

Add handlers for these three opcodes, so the specific engine has a
chance to handle them. Note that these three opcodes shouldn't normally
generate a response (that would indicate some kind of error), however
it is useful to have ep-engine handle the error as it has the
necessary context (DCP Stream etc) to print a useful error message.

Change-Id: Ifc841b0dcf70e7e1289dbd6019335ccbd1c5205b
Reviewed-on: http://review.couchbase.org/111653
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>