MB-34906: Make PDM seqnoAck highest persisted snapshotEnd
Problem:
The PDM seqnoAcks immediately if only Majority or
MajorityAndPersistToMaster prepares are encountered, but once a
PersistToMajority Prepare is found it must wait for the entire snapshot
to persist.

To do this, the PDM waits for the current persistedSeqno to exceed the
snapshotEnd, *but* the PDM only tracked the latest snapshotEnd seqno -
updated each time a snapshot end is received.

If a new snapshotEnd is received before persistence "catches up", the
PDM will wait until the end of the new snapshot. If this happens
continuously (i.e., under continuous load), the PDM might never ack.

Rejected Solution:
Add a "last prepare in snapshot" flag which is set on the last prepare
when notified of snapshotEnd. Rejected because persistence and
snapshotEnd notifications both call updateHighPreparedSeqno which would
have to scan the entire trackedWrites list to determine if a snapshot
had ended yet - this might lead to poor performance.

Solution:
Track a queue of snapshotEnd seqnos, and seqnoAck the highest
persisted, rather than waiting to persist the latest

Change-Id: Ib59149078e7b61dda20ac46174618a183adb4b61
Reviewed-on: http://review.couchbase.org/111642
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>