MB-34873: disk snapshots - Ack snapEnd seqno once persisted
Problem 1:
Any prepare received by a replica from a disk snapshot may have deduped
an earlier prepare of a higher level.

For example, the following ops (for the same key)

PRE(l=PersistToMajority) CMT PRE(l=Majority) CMT

May be deduped to

PRE(l=Majority) CMT

If we acked this prepare immediately, were we to (say) be promoted to
active and immediately crash, we may be left with no value on disk for
the key BUT the PersistToMajority op may have returned SUCCESS to the
client (it was committed) - we have broken the durability agreement for
that op.

Problem 2:
PRE(persistMajority), CMT, PRE(), ABORT, SET

may, after the abort has been purged be sent as:

SET

and we have no way of knowing a durability op was ever present.

Solution:
Advance the HPS to snapshotEndSeqno and seqnoAck once a full snapshot is
persisted, _just in case_ any prepares were deduped.

We could ack the latest prepare we are tracking, but we may be unaware
of some prepares due to Problem 2, so ack snapshotEndSeqno - any
prepares before this seqno are definitely prepared anyway.

Change-Id: I9822ea608da79e4ac55f8f4f42cabe545e26adb6
Reviewed-on: http://review.couchbase.org/111643
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>