MB-34879: Ensure correct datatype for uncompressed doc with xattrs
There is a bug that shows up on SDKs that support compression.  The
issue is with documents that have xattrs and are compressed.  With a get
request the document is uncompressed to allow the xattrs to be stripped
before the document is sent to the client.  However we do not clear the
snappy datatype on the document before sending the document.

On a client that does not support compression that is OK because we
set the document datatype based on the intersect of what the document
datatype is and what the client supports and hence the snappy datatype
is cleared.  However on a client that supports compression the snappy
datatype is not cleared.

This results in the server sending a document that is marked as snappy
compressed but the payload is not compressed.  Therefore the SDK raises
an error when it attempts to decompress the payload.

The fix is to clear the compression datatype on the document when it
is decompressed on the server, before the xattrs are stripped.

Change-Id: I90103a1ecb3de575d81c9f84181efaf044bd4a57
Reviewed-on: http://review.couchbase.org/111649
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>