Fix off-by-one error with removeClosedUnrefCheckpoints.limit
There is an off-by-one error in
CheckpointManager::removeClosedUnrefCheckpoints() when the `limit`
argumeent is used - the last unreferenced Checkpoint is not removed.

This is because the early `break` from the for-loop doesn't increment
`it` - as it would normally be in the for loop iteration expression -
note that `it` indicates the first checkpoint to keep, *not* the last
one to remove (as per std::list::splice taking the range [first, last)
(i.e. last is exclusive).

For example, consider 2 checkpoints where the only cursors are in the
final checkpoint:

    1:CLOSED, 2:OPEN
              ^
              |
              Persistence cursor

We should be able to remove Checkpoint 1 when limit==1 is specified,
however due to the aforementioned bug we don't remove any.

Change-Id: I355b5f953bb8e22993e86fff2a11f2794fa368e6
Reviewed-on: http://review.couchbase.org/111654
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>