MB-33805: Move memcached topkeys test to new testsuite
The current test is run after many write heavy test and is flawed
in the sense that it relies on topkeys returning the last writen
value to each shard. Move the topkeys tests to a new testsuite so
that they can be run in isolation (and a new file for clarity).

Change-Id: I001be4eddf674a174505b8602a5b812b26d85911
Reviewed-on: http://review.couchbase.org/111685
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>