MB-34920: Do not track prepare update as new prepare
Currently we track the update of a prepare as a new prepare
(incrementing the number of prepares on disk) which causes us
to overcount the number of prepares on disk. This results in an
incorrect document counter for full eviction post-warmup.

Correct the tracking of prepares on disk by moving the tracking
logic from EPBucket::flushVBucket to CouchKVStore::saveDocsCallback
as we know at this point if the item is new or updated.

Change-Id: I3246f69f6cc6368e98fad7b59f979428d6ac47b7
Reviewed-on: http://review.couchbase.org/111733
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>