MB-34780: Clear the pending sync writes during bucket shutdown
During bucket deletion all of the pending sync writes needs
to be terminated for bucket deletion to complete. As part
of bucket deletion we tear down all of the DCP streams so
there is no way for the ack/nack's from the replicas to
arrive and unblock the cookie and continue bucket deletion.

Change-Id: I1f2801c69cb1ee35cd0cfa4622d7ab5dc847f1e4
Reviewed-on: http://review.couchbase.org/111536
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>