MB-34951: Ignore StreamEnd if opaque doesn't match
The DCP option 'send_stream_end_on_client_close_stream' was added via
MB-26363 to have the Producer send a STREAM_END message even when the
Consumer initiated the close. This was requested by an external DCP
client (Go SDK), however the ep-engine DcpConsumer _also_ requested
this feature.

However this can cause (benign) warnings during rebalance when the
initial (backfill) Stream is closed and the takeover stream
created. The problem is the following sequence of messages (all for
the same vbid):

    DcpConsumer            ns_server proxy                DcpProducer

    <--- 1. CloseStream Req ------
    ---- 2. CloseStream Resp ---->
                                       ---- 3. CloseStream Req --->
                                       <--- 4. CloseStream Resp ---

    <--- 5. AddStream(Takeover) --
    <--- 6. StreamEnd Request -------------------------------------
    ---- 7. StreamEnd Response [KEY_ENOENT] ---------------------->

    ---- 8. StreamRequest ---------------------------------------->

ns_server is not aware of `send_stream_end_on_client_close_stream`; so
it doesn't wait for `6. StreamEnd Request` from the Producer before
sending the AddStream request to the Consumer. As such, the Consumer
has already removed the entry in it's streams map for the given
vbucket, so when it receives the StreamEnd request from the Producer
it returns KEY_ENOENT error as the (old) opaque doesn't match the
(new) opaque for the takeover stream.

send_stream_end_on_client_close_stream is useful for ep-engine to have
an explicit window on when potential mismatched opaques could appear
(after the SteamEnd request no more should be sent from the Producer)
- and indeed it's presence is required to address MB-34850. As such we
cannot simply disable send_stream_end_on_client_close_stream.

The alternative is to simply ignore the StreamEnd if the given opaque
no longer matches (i.e. because another stream has already been
started).

Change-Id: Ib6baf414aa5b4c306f9146818db2d25ac4326f4f
Reviewed-on: http://review.couchbase.org/111746
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>