MB-34946: Use HashTable::FindCommitResult in processSoftDelete
In a following change, we need to make use of the FindCommitResult
in VBucket::processSoftDelete as it contains both the pending and
committed StoredValues. To keep a single processSoftDelete
interface, update all callers to pass a FindCommitResult instead of
a single StoredValue&. Also, add a processSoftDeleteInner function
as we will need to modify which StoredValue& v is deleted but
assignment is not allowed.

Change-Id: I3b7060a111c760a2194b8f5abac6344f5148b656
Reviewed-on: http://review.couchbase.org/111858
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>