MB-34996: Do not abort prepares with infinite timeout
If we warmup or promote a replica to active then we create
SyncWrite objects with an infinite timeout as these prepares
MUST be committed as they could have been committed pre-warmup
or by the old active. Currently, the code will attempt to abort
these prepares if a subsequent topology change makes durability
impossible. This breaks durability though as we MUST commit these
prepares.

Update the code to never attempt to abort a SyncWrite with an
infinite timeout.

Change-Id: I206ec581e827f1ca12ff5637c0e2fb9aabac4284
Reviewed-on: http://review.couchbase.org/111861
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>