MB-34910: Track HPS/HCS as member variables of the ADM::State
Instead of deriving the HPS and HCS when they are read, keep
a tracked/cached counter in the ADM::State which is simply
returned when needed.

This is also required to simplify the warmup of the ADM as
we can just assign the pre-warmup value.

Note: highPreparedSeqno is tracked as weak monotonic counter
becase updateHighPreparedSeqno can run and make no change,
rather than detect when no-change occured we just tolerate
assignment of the current value.

Change-Id: I914637548fdfecd419d6492561a4a3319fe88b26
Reviewed-on: http://review.couchbase.org/111863
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>