MB-34910: Warmup DurabilityMonitor
Following a restart the DurabilityMonitor should be in a logically
equivalent state to it's pre-warmup state. Of interest in this patch are
the restoration of

1. High Prepared Seqno
2. High Completed Seqno
3. Highest Tracked Seqno (ActiveDurabilityMonitor only)

To achieve the restoration of 1 and 2 and with inclusion of the previous
patch warmup can just reset the variables to their pre-warmup values.

For the restoration of 3, that is just assigned the value of 1.

Further to these changes extension of the existing warmup testing showed
that following warmup, any loaded prepares lose the 'ack' count for the
current node, all prepares found on disk must have at least an 'ack'
from the node itself. This is addressed when constructing the ADM by
calling updateHighPreparedSeqno after loading all tracked writes.

Change-Id: If3033397a331bbcef06b0fe22d93185ba9993489
Reviewed-on: http://review.couchbase.org/111701
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>