MB-34453: Block SyncWrites if >2 replicas in chain
Due to issues with failover+rollback (see MB-34453), Sync Writes
cannot be supported with 2 auto failovers. The first auto-failover is
fine and hence we can support up to 2 replicas safely without losing
any durably written data – this is obviously because with 2 replicas
users would expect only 1 auto-failover safely. The problem arises
when you have 3 replicas.

As such, disallow use of 3 replicas if Sync Writes are used. If a
client attempts to perform a SyncWrite with >2 replicas, the request
will fail with EDurabilityImpossible.

The behaviour is controlled via a new bucket-level configuration paramter:

- sync_writes_max_allowed_replicas: The maximum number of supported
  replicas for SyncWrites. Attempts to issue SyncWrites against a
  topology with more replicas than this setting will fail with
  DurabilityImpossible.

The default value for this setting is 2, however it _could_ be
overridden at bucket creation time if necessary.

Change-Id: Ifb117bff7c8f5f4afb9cdfd24a5b1538efd36ab3
Reviewed-on: http://review.couchbase.org/111919
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>