MB-35062: Tolerate prepare(k), delete(k) in a disk snapshot
When receiving a disk snapshot the replica must allow both
items in and store both to disk. This change updates processSoftDelete
so that the delete is not considered a "no-key" error.

The orginal defect occurred because the following occurs on
the active:

    1) Prepare(k)
    2) Commit(k)

Disk now looks like:

    seq:1 prepare(k)
    seq:2 commit(k)

Next a delete(k) occurs and disk now looks like:

    seq:1 prepare(k)
    seq:3 delete(k)

If we now build a replica from this disk sequence, DCP will transmit:

    seq:1 DCP prepare(k)
    seq:3 DCP delete(k)

The replica must process both of these messages and persist both a
prepare and delete, prior to this patch the delete was dropped
because the vbucket had no committed value to delete.

Change-Id: I4f93c8e9c1033e3e53fd57e9ec8e9db5545a7047
Reviewed-on: http://review.couchbase.org/112143
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>