MB-35133: Test HashTable SV state post-rollback
To verify that all StoredValues in the HashTable are correct
post-rollback, add a dumpStoredValuesAsJson function to HashTable
which we can use to compare the HashTable state before and after
a rollback.

This patch updates all non-durability tests to use this function,
but leaves the durability tests as they are as they are currently
all disabled.

Note, we actually compare the json strings produced by
nlohmann::json::dump() here as there is a bug in our version of GTest
that causes a stack overflow when attempting to print the
nlohmann::json object. Additionally, by passing in the 0 parameter,
nlohmann::json::dump() will split our objects with newlines which
GTest can utilise to give us a diff if the string comparison fails.

Change-Id: I69204be1d57e0250b7c7b07658ad25fd2b90da89
Reviewed-on: http://review.couchbase.org/112150
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>